Added sorting and improved code

+++ router/main.js
@@ -1,5 +1,6 @@
 const moment = require('moment');
-module.exports = function(app, todos, fs, database, crypto)
+
+module.exports = function(app, todos, fs, database, crypto, sortvar)
 {
     //
     // list the todos
@@ -14,7 +15,7 @@ module.exports = function(app, todos, fs, database, crypto)
 
             // get the count of non-empty elements  
             var count = todos.filter(function(value) {return value !== undefined}).length;
-
+            
             if(count === 0){
                 req.flash('empty', 'Please add your first to do.');
             }
@@ -23,7 +24,7 @@ module.exports = function(app, todos, fs, database, crypto)
             }
             
             // render the page
-            res.render('index', {message: req.flash('empty'), moment: moment, ptitle:"Todos", todos: todos, count:count, username: req.session.username, userid: req.session.userid});
+            res.render('index', {message: req.flash('empty'), sort: sortvar, moment: moment, ptitle:"Todos", todos: todos, count:count, username: req.session.username, userid: req.session.userid});
         }
         else res.redirect('/login');
     });
@@ -33,34 +34,47 @@ module.exports = function(app, todos, fs, database, crypto)
     //
     app.get('/add', function(req, res){
         if(req.session.email) {
+            uid = req.session.userid;
+            database.query("SELECT * FROM todos WHERE userid = '"+uid+"'", function(error, rows){
+                if(error){console.log('get-todos error: ', error);}
+                else{
+                    // load todos from the database
+                    rows.forEach(function(data){
+                        todos[data.id] = data;
+                    })
+                }
+            });
             res.render('add', {ptitle:"Add todo", username: req.session.username, email: req.session.email, userid: req.session.userid});
         }
         else res.redirect('/login');
     });
 
     app.get('/add-submit', function(req, res){
-        let todotext = req.query.todo;
-        let priority = req.query.priority;
-        let category = req.query.category;
-        let notes = req.query.notes;
-        let userid = req.query.userid;
+        if(req.session.email) {
+            let todotext = req.query.todo;
+            let priority = req.query.priority;
+            let category = req.query.category;
+            let notes = req.query.notes;
+            let userid = req.query.userid;
 
-        // save todo to the database
-        database.query(
-        `INSERT INTO todos (todo, priority, category, notes, userid) VALUES ('${todotext}', '${priority}', '${category}', '${notes}', '${userid}')`,
-        function (error, results, fields){
-            if(error) throw console.log('add-submit error: ', error);
-            
-            // res.send(results);
-            // console.log("results " + results)
-            else{
-                // save todo to the array
-                todos.push(req.query.todo);
-            
-                // redirect to get todos for updated todos
-                res.redirect('/get-todos');
-            }
-        });
+            // save todo to the database
+            database.query(
+            `INSERT INTO todos (todo, priority, category, notes, userid) VALUES ('${todotext}', '${priority}', '${category}', '${notes}', '${userid}')`,
+            function (error, results, fields){
+                if(error) throw console.log('add-submit error: ', error);
+                
+                // res.send(results);
+                // console.log("results " + results)
+                else{
+                    // save todo to the array
+                    todos.push(req.query.todo);
+                
+                    // redirect to get todos for updated todos
+                    res.redirect('/get-todos');
+                }
+            });
+        }
+        else res.redirect('/login');
     });
 
     // 
@@ -68,42 +82,57 @@ module.exports = function(app, todos, fs, database, crypto)
     //
     app.get('/update', function(req, res){
         if(req.session.email) {
-            res.render('update', {moment: moment, ptitle:"Update Todos", todos: todos, username: req.session.username, email: req.session.email, userid: req.session.userid});
+            uid = req.session.userid;
+            todos = [];
+            // console.log("get todos userid " + uid);
+            database.query(`SELECT * FROM todos WHERE userid = ${uid}`, function(error, rows){
+                if(error){console.log('get-todos error: ', error);}
+                else{
+                    // load todos from the database
+                    rows.forEach(function(data){
+                        todos[data.id] = data;
+                    })
+                    res.render('update', {moment: moment, ptitle:"Update Todos", todos: todos, username: req.session.username, email: req.session.email, userid: req.session.userid});
+                }
+            });
         }
         else res.redirect('/login');
         
     });
 
     app.get('/update-submit', function(req, res){
-        // update todo fields
-        // console.log("priority field " + req.query.priority)
-        let todoid = req.query.todoid;
-        let priority = req.query.priority;
-        let cat = req.query.category;
-        let notes = req.query.notes;
-        let todo = req.query.todo;
+        if(req.session.email) {
+            // update todo fields
+            // console.log("priority field " + req.query.priority)
+            let todoid = req.query.todoid;
+            let priority = req.query.priority;
+            let cat = req.query.category;
+            let notes = req.query.notes;
+            let todo = req.query.todo;
 
-        var sql = "UPDATE todos SET category ='"+cat+"', priority ='"+priority+"', notes ='"+notes+"', todo ='"+todo+"' WHERE id='"+todoid+"'";
-        
-        database.query(sql, function(err, rows){
-            if(err) {
-                return console.log(err);
-            }
-            else
-            {
-                // redirect to get todos for updated todos
-                res.redirect('/get-todos');
-            }
-        });
+            var sql = "UPDATE todos SET category ='"+cat+"', priority ='"+priority+"', notes ='"+notes+"', todo ='"+todo+"' WHERE id='"+todoid+"'";
+            
+            database.query(sql, function(err, rows){
+                if(err) {
+                    return console.log(err);
+                }
+                else
+                {
+                    // redirect to get todos for updated todos
+                    res.redirect('/get-todos');
+                }
+            });
+        }
+        else res.redirect('/login');
     });
     //
     // completed checkbox
     //
     app.post('/checkcompleted', function(req, res){
-        let id = req.body.tid;
-        let completed = req.body.checkval;
-        
         if(req.session.email) {
+            let id = req.body.tid;
+            let completed = req.body.checkval;
+        
             // console.log(">>> id " + id + " completed " + completed);
             database.query("UPDATE `todos` SET `completed` ='"+completed+"' WHERE `id` ='"+id+"'",
             function (error, results, fields){
@@ -119,28 +148,42 @@ module.exports = function(app, todos, fs, database, crypto)
     //
     app.get('/del', function(req, res){
         if(req.session.email) {
-            res.render('delete', {moment: moment, ptitle:"Update Todos", todos: todos, username: req.session.username, email: req.session.email, userid: req.session.userid});
+            uid = req.session.userid;
+            todos = [];
+            // console.log("get todos userid " + uid);
+            database.query(`SELECT * FROM todos WHERE userid = ${uid}`, function(error, rows){
+                if(error){console.log('get-todos error: ', error);}
+                else{
+                    // load todos from the database
+                    rows.forEach(function(data){
+                        todos[data.id] = data;
+                    })
+
+                    res.render('delete', {moment: moment, ptitle:"Delete Todos", todos: todos, username: req.session.username, email: req.session.email, userid: req.session.userid});
+                }
+            });
         }
         else res.redirect('/login');
     });
 
     app.get('/del-submit/:id', function(req, res){
-        // delete from array
-        let id = req.params.id;
+        if(req.session.email) {
+            // delete from array
+            let id = req.params.id;
 
-        // console.log("delete todo " + id);
-
-        // delete todo from the database
-        database.query(
-            `DELETE FROM todos WHERE id = ${id}`,
-            function (error, results, fields){
-                if(error) {console.log('delete error: ', error);}
-                else{
-                    // redirect to get todos for updated todos
-                    res.redirect('/get-todos');
-                }
-        });
-
+            // console.log("delete todo " + id);
+            // delete todo from the database
+            database.query(
+                `DELETE FROM todos WHERE id = ${id}`,
+                function (error, results, fields){
+                    if(error) {console.log('delete error: ', error);}
+                    else{
+                        // redirect to get todos for updated todos
+                        res.redirect('/get-todos');
+                    }
+            });
+        }
+        else res.redirect('/login');
     });
 
     //
@@ -214,54 +257,158 @@ module.exports = function(app, todos, fs, database, crypto)
         });
     });
 
-    app.get('/logout-submit', function(req, res){
-        req.session.email = false;
-        res.redirect('/get-all-todos');
+    
+    //
+    // sorting
+    //
+
+    app.get('/priority/:pri', function(req, res){
+        if(req.session.email) {
+            uid = req.session.userid;
+            let priority = req.params.pri
+            //set sort so when a sort selection is made, css shows the selection as curent
+            sortvar = priority;
+            todos = [];
+            // console.log("get todos userid " + uid); 
+            database.query("SELECT * FROM todos WHERE userid = '"+uid+"' AND priority = '"+priority+"'", function(error, rows){
+                if(error){console.log('get-todos error: ', error);}
+                else{
+                    // load todos from the database
+                    rows.forEach(function(data){
+                        todos[data.id] = data;
+                    })
+                    
+                    // redirect to home
+                    res.redirect('/');
+                }
+            });
+        }
+        else res.redirect('/login');
     });
 
+    
+    app.get('/status/:stat', function(req, res){
+        if(req.session.email) {
+            uid = req.session.userid;
+            
+            if(req.params.stat === 'completed'){
+                sortvar = '1';
+            }
+            else{
+                sortvar = '0';
+            }
+            // console.log("sort: " + sortvar);
+            //set sort so when a sort selection is made, css shows the selection as curent
+            todos = [];
+            // console.log("get todos userid " + uid); 
+            database.query(`SELECT * FROM todos WHERE userid = ${uid} AND completed = ${sortvar}`, function(error, rows){
+                if(error){console.log('get-todos error: ', error);}
+                else{
+                    // load todos from the database
+                    rows.forEach(function(data){
+                        todos[data.id] = data;
+                    })
+                    
+                    // redirect to home
+                    res.redirect('/');
+                }
+            });
+        }
+        else res.redirect('/login');
+    });
+
+    app.get('/date/:event', function(req, res){
+        if(req.session.email) {
+            uid = req.session.userid;
+            
+            if(req.params.event === 'old'){
+                sortvar = 'ASC';
+            }
+            else{
+                sortvar = 'DESC';
+            }
+            console.log("sort: " + sortvar);
+            //set sort so when a sort selection is made, css shows the selection as curent
+            todos = [];
+            
+            database.query("SELECT * FROM todos WHERE userid = "+uid+" ORDER BY id "+sortvar, function(error, rows){
+                if(error){console.log('get-todos error: ', error);}
+                else{
+                    // load todos from the database
+                    rows.forEach(function(data, index){
+                        todos[index] = data;
+                        // console.log("record >>> " + JSON.stringify(data) + " dataid: " + index);
+                    })
+
+                    // redirect to home
+                    res.redirect('/');
+                    
+                }
+            });
+        }
+        else res.redirect('/login');
+    });
+    
     //
-    // ajax calls
+    // Get Todos
     //
+
 
     //grab all records that match the signed in user and redirect to home
     app.get('/get-todos', function(req, res){
-        uid = req.session.userid;
-        todos = [];
-        // console.log("get todos userid " + uid);
-        database.query(`SELECT * FROM todos WHERE userid = ${uid}`, function(error, rows){
-            if(error){console.log('get-todos error: ', error);}
-            else{
-                // load todos from the database
-                rows.forEach(function(data){
-                    todos[data.id] = data;
-                })
+        if(req.session.email) {
+            uid = req.session.userid;
 
-                // redirect to home
-                res.redirect('/');
-            }
-        }); 
+            //reset sort so when home is clicked css shows home as curent
+            sortvar = "";
+            todos = [];
+
+            // console.log("get todos userid " + uid);
+            database.query(`SELECT * FROM todos WHERE userid = ${uid}`, function(error, rows){
+                if(error){console.log('get-todos error: ', error);}
+                else{
+                    // load todos from the database
+                    rows.forEach(function(data){
+                        todos[data.id] = data;
+                    })
+
+                    // redirect to home
+                    res.redirect('/');
+                }
+            });
+        }
+        else res.redirect('/login');
     });
 
     //a type of refresh, otherwise server must be restarted to see changes
     //grab all records and redirect user to login if signed out or home screen if signed in
     app.get('/get-all-todos', function(req, res){
-        todos = [];
-        database.query(`SELECT * FROM todos`, function(error, rows){
-            if(error){console.log('get-all-todos error: ', error);}
-            else{
-                // load todos from the database
-                rows.forEach(function(data){
-                    todos[data.id] = data;
-                })
-                if(req.session.email) {
-                    // redirect to login
-                    res.redirect('/');
-                }
+        if(req.session.email) {
+            todos = [];
+            database.query(`SELECT * FROM todos`, function(error, rows){
+                if(error){console.log('get-all-todos error: ', error);}
                 else{
-                    // redirect to login
-                    res.redirect('/login');
-                } 
-            }
-        }); 
+                    // load todos from the database
+                    rows.forEach(function(data){
+                        todos[data.id] = data;
+                    })
+                    if(req.session.email) {
+                        // redirect to login
+                        res.redirect('/');
+                    }
+                    else{
+                        // redirect to login
+                        res.redirect('/login');
+                    } 
+                }
+            }); 
+        }
+        else res.redirect('/login');
     });
+
+    app.get('/logout-submit', function(req, res){
+        req.session.email = false;
+        res.redirect('/get-all-todos');
+    });
+
 }
